pyproject.toml
[project]
name = "staffingagent-reconciliation"
version = "0.1.0"
description = "VMS vs ATS reconciliation pipeline"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
  "anthropic>=0.40.0",
  "langgraph>=0.2.0",
  "langchain>=0.3.0",
  "langchain-anthropic>=0.3.0",
  "langsmith>=0.1.0",
  "pandas>=2.0.0",
  "rapidfuzz>=3.0.0",
  "python-dotenv>=1.0.0",
]

[tool.pyright]
pythonVersion = "3.11"
include = ["src", "tests"]
extraPaths = ["."]
